fix(fx-lib): .5K is 5512 Hz and the HUD tapes are 32-bit (#491)
Two of the seven bugs the #482 audit found, both still live, both user-visible
corruption, and both invisible to a round-trip test by construction.

.5K SAMPLE RATE — audio.cpp returned 5000; the rate is 5512 (11025/2). 781 files,
every one of them 10% long. The rate is never stored in the file (it comes from the
extension), so NO repack can contradict it -- and test_audio.cpp asserted 5000, so
the suite agreed with the bug. A test that asserts what the code does is not a test.

The briefing videos are an independent clock, and they settle it: audio is shared
per 3-character group, the paired .FBC holds one u32 per frame, the .VDO header
gives the fps. IQC.5K is 619,237 bytes against 1,685 frames at 15 fps = 112.33 s:

    at 5000 Hz -> 123.85 s   (+10.2%)
    at 5512 Hz -> 112.34 s   ( +0.0%)

tests/test_audio.cpp now reconciles all 106 tracks in the install against that clock.
The rule is an inequality, not an equality: a narration may end before its clip does
(ZAC is 8.0 s of speech over a 14.7 s clip -- the one such track), but it may never
OUTRUN the video, which is exactly the shape of a rate that is too low. The median
ratio pins the other side. Verified to FAIL at 5000 and pass at 5512.

HUD TAPE WIDTHS — the 11 heading/speed/altitude tape fields were typed S16; they are
i32. The table's own stride says so (0x1E1 -> 0x1E5 -> 0x1E9 -> 0x1ED is four apart,
where the flight-path-marker fields below are two), and across all 46 shipped HUDs
every one is sign-extended over four bytes, 44 of them carrying a negative.

Reading LOOKED right, because the low half is the value. Writing corrupted:
`fx hud set A7.HUD speed_tape.dx=40` wrote `28 00` and left the old `ff ff` behind,
so the engine read -65496. An unedited repack rewrites the bytes it just read, so
the census could never see it.

tests/test_hud.cpp now writes every field the editor can write, on all 46 install
HUDs, with the sign flipped -- and asserts the bytes on disk, not just the
round-trip. Verified to fail (509 assertions) against the old 16-bit write.

The synthetic fixture had encoded the same mistake (put16 on a 4-byte field) and was
defending it; it now matches the real files. HudParam::value widens to int32_t.

Docs: HUD.md said "all gauge positions are stored as signed s16" -- the sentence the
codec was written from -- and now carries the two widths with the evidence. 11K.md
already had 5512 right and now shows the derivation. VDO.md claimed the IQC group is
"silent"; it is not -- its narration is the .5K, the very file that proves the rate.

+++ lib/include/fx/hud.h
@@ -13,7 +13,10 @@ namespace fx {
 struct HudParam {
     std::string gauge;
     std::string field;
-    int16_t     value;
+    // The tape gauges store their parameters as 32-bit, sign-extended (verified across
+    // all 46 shipped HUDs); the rest are 16- or 8-bit. A 16-bit value cannot hold them,
+    // and writing one back 16 bits wide left the high half stale — see hud.cpp.
+    int32_t     value;
 };
 
 struct HudFile {
